+++ resources/ReleaseNotes.md
@@ -1,2 +1,3 @@
+- Add calibration data files
 - Add utility method to retrieve specific HK type from vector of HK
 - Fix issue with app icons not founds